+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `Text` class to create Text Elements.
 - Backticks (`` ` ``) can now be escaped in multiline codes.
+- Footnotes are implemented.
 
 
 ## 2.0.0 - 2020-08-21